+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - [#340](https://github.com/estruyf/vscode-demo-time/issues/340): Fix an issue
   with the `path` label for the `executeScript` action in the step editor
+- [#341](https://github.com/estruyf/vscode-demo-time/issues/341): Added the
+  `args` property to the `executeScript` action to allow passing command-line
+  arguments to the script being executed
 
 ## [1.12.0] - 2025-12-16